\name{HTMLColumns}
\alias{HTMLColumns}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLColumns(Ncol=2, Arg=list(c(), c()), File="", Align="center", Border=0, Width="100\%")
~move the above line to just above the first optional argument
}
\arguments{
\item{Ncol}{
~Describe Ncol here
}
\item{Arg}{
~Describe Arg here
}
\item{File}{
~Describe File here
}
\item{Align}{
~Describe Align here
}
\item{Border}{
~Describe Border here
}
\item{Width}{
~Describe Width here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
