\name{HTMLCommand}
\alias{HTMLCommand}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLCommand(text, File, Size="+0", Num="")
~move the above line to just above the first optional argument
}
\arguments{
\item{text}{
~Describe text here
}
\item{File}{
~Describe File here
}
\item{Size}{
~Describe Size here
}
\item{Num}{
~Describe Num here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
