\name{HTMLDemo}
\alias{HTMLDemo}
\title{
Demo of the package main functions
}
\description{
Compute an HTML document called R2HTMLDemo.html using some functions of the package. It's a good introduction for the new user.
}
\usage{
HTMLDemo(path="./")
}
\arguments{
\item{path}{
the path to the directory where the generated HTML document will be stored.
}}
\examples{
HTMLDemo()
}
\keyword{demo}

