\name{HTMLExport.anova}
\alias{HTMLExport.anova}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.anova(x, File="", digits=.Options$digits, quote=F, drop=F, FontName="Arial", Align="left", Bold=T, \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{digits}{
~Describe digits here
}
\item{quote}{
~Describe quote here
}
\item{drop}{
~Describe drop here
}
\item{FontName}{
~Describe FontName here
}
\item{Align}{
~Describe Align here
}
\item{Bold}{
~Describe Bold here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
