\name{HTMLExport.array}
\alias{HTMLExport.array}
\title{
Use ´HTMLExport' on a array object
}
\description{
Method for the function HTMLExport() for objects inheriting from class array
}
\usage{
HTMLExport.array(x, File="", \dots)
}
\value{
This is a method for the function HTMLExport() for objects inheriting from class array.
See HTMLExport or HTMLExport.default for the general behavior of this function
and for the interpretation of x.
}
\seealso{
´HTMLExport'
}
\keyword{array}
% Converted by Sd2Rd version 1.10.
