\name{HTMLExport.atomic}
\alias{HTMLExport.atomic}
\title{
Use ´HTMLExport' on data with atomic modes.
}
\description{
Method for the function HTMLExport() for objects of class NULL
}
\usage{
HTMLExport.atomic(x, File="", \dots)
}
\value{
This is a method for the function HTMLExport() for objects of class NULL (atomic modes).
See HTMLExport or HTMLExport.default for the general behavior of this function
and for the interpretation of x.
}
\seealso{
´HTMLExport'
}
\keyword{atomic}
% Converted by Sd2Rd version 1.10.
