\name{HTMLExport.by}
\alias{HTMLExport.by}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.by(x, \dots, vsep="<HR WIDTH=90\% HEIGHT=1>")
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{...}{
~Describe \dots{} here
}
\item{vsep}{
~Describe vsep here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
