\name{HTMLExport.data.frame}
\alias{HTMLExport.data.frame}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.data.frame(x, File=HTML.File, Border=0, TextSize="-1", FontName="times", FirstRowBold=F, FirstColumnBold=F, ColorFirstRow="E4E4E4", ColorFirstColumn="C0C0C0", ColorCellInside="FFFFFF", Align="right", Alignx="center", Bold=F)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{Border}{
~Describe Border here
}
\item{TextSize}{
~Describe TextSize here
}
\item{FontName}{
~Describe FontName here
}
\item{FirstRowBold}{
~Describe FirstRowBold here
}
\item{FirstColumnBold}{
~Describe FirstColumnBold here
}
\item{ColorFirstRow}{
~Describe ColorFirstRow here
}
\item{ColorFirstColumn}{
~Describe ColorFirstColumn here
}
\item{ColorCellInside}{
~Describe ColorCellInside here
}
\item{Align}{
~Describe Align here
}
\item{Alignx}{
~Describe Alignx here
}
\item{Bold}{
~Describe Bold here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
