\name{HTMLExport.graphics}
\alias{HTMLExport.graphics}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.graphics(x, File="", GraphDirectory="c:\\TEMP\\", GraphFileName="", GraphSaveAs="JPG", GraphBorder=1, Align="center", \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{GraphDirectory}{
~Describe GraphDirectory here
}
\item{GraphFileName}{
~Describe GraphFileName here
}
\item{GraphSaveAs}{
~Describe GraphSaveAs here
}
\item{GraphBorder}{
~Describe GraphBorder here
}
\item{Align}{
~Describe Align here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
