\name{HTMLExport.lm}
\alias{HTMLExport.lm}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.lm(lmobject, File="", Directory=".\\", Desc=F, \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{lmobject}{
~Describe lmobject here
}
\item{File}{
~Describe File here
}
\item{Directory}{
~Describe Directory here
}
\item{Desc}{
~Describe Desc here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
