\name{HTMLExport.summary.lm}
\alias{HTMLExport.summary.lm}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLExport.summary.lm(x, File="", digits=max(3, .Options$digits - 3), \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{digits}{
~Describe digits here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
