\name{HTMLExport.title}
\alias{HTMLExport.title}
\title{
Writes a title to an HTML page
}
\description{
~brief description
}
\usage{
HTMLExport.title(x, File=HTML.File, \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
´as.title'
~put functions to SEE ALSO here
}
\examples{
}
\keyword{title}
% Converted by Sd2Rd version 1.10.
