\name{HTMLInsertBreak}
\alias{HTMLInsertBreak}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLInsertBreak(File="", Width="100\%", Align="center", Size="3")
~move the above line to just above the first optional argument
}
\arguments{
\item{File}{
~Describe File here
}
\item{Width}{
~Describe Width here
}
\item{Align}{
~Describe Align here
}
\item{Size}{
~Describe Size here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
