\name{HTMLPlot.lm}
\alias{HTMLPlot.lm}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLPlot.lm(x, File="", residuals=NULL, smooths=F, rugplot=F, id.n=3, ask=F, which.plots=NULL, GraphDirectory, Name="reg", \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{x}{
~Describe x here
}
\item{File}{
~Describe File here
}
\item{residuals}{
~Describe residuals here
}
\item{smooths}{
~Describe smooths here
}
\item{rugplot}{
~Describe rugplot here
}
\item{id.n}{
~Describe id.n here
}
\item{ask}{
~Describe ask here
}
\item{which.plots}{
~Describe which.plots here
}
\item{GraphDirectory}{
~Describe GraphDirectory here
}
\item{Name}{
~Describe Name here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
