\name{HTMLSink}
\alias{HTMLSink}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLSink(value=T, HTMLOutputFile=".\\output.html", HTMLOutputIndex=".\\index.html", HTMLOutputList=".\\list.html", HTMLOutputTitle=".\\high.html")
~move the above line to just above the first optional argument
}
\arguments{
\item{value}{
~Describe value here
}
\item{HTMLOutputFile}{
~Describe HTMLOutputFile here
}
\item{HTMLOutputIndex}{
~Describe HTMLOutputIndex here
}
\item{HTMLOutputList}{
~Describe HTMLOutputList here
}
\item{HTMLOutputTitle}{
~Describe HTMLOutputTitle here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
