\name{HTMLTitre}
\alias{HTMLTitre}
\title{
~function to do ???
}
\description{
~brief description
}
\usage{
HTMLTitre(Titre, Border=0, BackColor="C0C0C0", Size="+2", Color="000063", Width="100\%", File=HTML.File, Align="CENTER", FontName="Arial", \dots)
~move the above line to just above the first optional argument
}
\arguments{
\item{Titre}{
~Describe Titre here
}
\item{Border}{
~Describe Border here
}
\item{BackColor}{
~Describe BackColor here
}
\item{Size}{
~Describe Size here
}
\item{Color}{
~Describe Color here
}
\item{Width}{
~Describe Width here
}
\item{File}{
~Describe File here
}
\item{Align}{
~Describe Align here
}
\item{FontName}{
~Describe FontName here
}
\item{...}{
~Describe \dots{} here
}}
\value{
~Describe the value returned
}
\section{Side Effects}{
~describe any side effects if they exist
}
\details{
~explain details here.
}
\references{
~put references here, make other sections like NOTE and WARNING with .SH
}
\seealso{
~put functions to SEE ALSO here
}
\examples{
}
\keyword{~keyword}
% Converted by Sd2Rd version 1.10.
