\name{HTML.BR}
\alias{HTML.BR}
\title{
Break Return
}
\description{
Insert a break return to the output (go to the line).
}
\usage{
HTML.BR(File=HTML.File)
}
\arguments{
\item{File}{
HTML file to write the HTML code <BR>.
By default, HTML.File is the file used by S2HTML for output.
If equal to "", then output is redirected to SPlus screen.
}}
\value{
None.
}
\section{Side Effects}{
Add the <BR> code to the HTML output.
}
\seealso{
´HTML.HR' (separation line)
}
\examples{
HTML.BR()
}
\keyword{return}
% Converted by Sd2Rd version 1.10.
