\name{HTLMReplaceNA}
\alias{HTMLReplaceNA}
\alias{HTMLCommand}
\title{Internal R2HTML functions}
\description{   Internal R2HTML functions}
\usage{
HTMLReplaceNA(Vec, Replace = " ")
HTMLCommand(x,File=.HTML.File,Num="",menu=F,target="index_main.html")
}
\arguments{
  \item{Vec}{ string }
  \item{Replace}{ string to use for missing values }
  \item{x}{ a string corresponding to a R command}
  \item{File}{ the target HTML file }
  \item{Num}{ number of the command}
  \item{menu}{}
  \item{target}{}
}
\details{
     These are not to be called by the user.
}
\examples{

}
\keyword{misc}
